catalog/callbacks/finops_after_agent/requirements.txt
google-adk>=1.9.0
google-cloud-bigquery>=3.35.1
